Implement Authd tests code improvements.
