+++ deps/wazuh_testing/wazuh_testing/authd.py
@@ -7,10 +7,7 @@ copyright:
 '''
 
 import re
-import pytest
-import time
-from wazuh_testing.tools import CLIENT_KEYS_PATH, LOG_FILE_PATH
-from wazuh_testing.wazuh_db import query_wdb
+from wazuh_testing.tools import LOG_FILE_PATH
 from wazuh_testing.tools.monitoring import FileMonitor, make_callback, AUTHD_DETECTOR_PREFIX
 
 
@@ -25,11 +22,11 @@ def create_authd_request(input):
     Args:
         input (dict): Dictionary with the content of the request command.
     """
-    command = ""
+    command = ''
 
     if 'password' in input:
         password = input['password']
-        command = command + f'OSSEC PASS: {password} '
+        command = command + f"OSSEC PASS: {password} "
 
     command = command + 'OSSEC'
 
@@ -54,8 +51,14 @@ def create_authd_request(input):
     return command
 
 
-# Functions
 def validate_authd_logs(expected_logs, log_monitor=None):
+    """
+    Validates if a set of expected messages were written in the logs.
+
+    Args:
+        expected_logs (list): List with all the logs to be found.
+        log_monitor (FileMonitor): File monitor object to be used to find logs. If None, a new instance is created.
+    """
     if not log_monitor:
         log_monitor = FileMonitor(LOG_FILE_PATH)
 
@@ -66,132 +69,46 @@ def validate_authd_logs(expected_logs, log_monitor=None):
                           error_message=f"Expected log does not occured: '{log}'")
 
 
-def validate_argument(received, expected, argument_name):
-    if received != expected:
-        return 'error', f"Invalid '{argument_name}': '{received}' received, '{expected}' expected."
-    else:
-        return 'success', ''
+def parse_authd_response(response):
+    """
+    Parses an Authd response into a dictionary.
+
+    Args:
+        response (str): Raw response received from Authd.
+    Returns:
+        dict: Parsed components of the response.
+    """
+    response_dict = {}
+    try:
+        header, payload = response.split(sep=' ', maxsplit=1)
+        if header == 'OSSEC':
+            response_dict['status'] = 'success'
+            # The agent key is sent within '' and each component is separated by white spaces:
+            # K:'001 agent_name any TopSecret
+            agent_key = payload.split('\'')[1]
+            response_dict['id'], response_dict['name'], response_dict['ip'], response_dict['key'] = agent_key.split(' ')
+        elif header == 'ERROR:':
+            response_dict['status'] = 'error'
+            response_dict['message'] = payload
+        else:
+            raise
+
+    except Exception:
+        raise IndexError(f"Authd response does not have the expected format: '{response}'")
+    return response_dict
 
 
 def validate_authd_response(response, expected):
     """
     Validates if the different items of an Authd response are as expected. Any item inexistent in expected won't
     be validated.
-
     Args:
         response (str): The Authd response to be validated.
         expected (dict): Dictionary with the items to validate.
     """
-    response = response.split(sep=" ", maxsplit=1)
-    status = response[0]
-    result = 'success'
-    err_msg = ''
-    if expected['status'] == 'success':
-        result, err_msg = validate_argument(status, 'OSSEC', 'status')
-        if result != 'success':
-            return result, err_msg
+    response_dict = parse_authd_response(response)
 
-        agent_key = response[1].split('\'')[1::2][0].split()
-        id = agent_key[0]
-        name = agent_key[1]
-        ip = agent_key[2]
-        key = agent_key[3]
-
-        if 'id' in expected:
-            result, err_msg = validate_argument(id, expected['id'], 'id')
-            if result != 'success':
-                return result, err_msg
-
-        if 'name' in expected:
-            result, err_msg = validate_argument(name, expected['name'], 'name')
-            if result != 'success':
-                return result, err_msg
-
-        if 'ip' in expected:
-            result, err_msg = validate_argument(ip, expected['ip'], 'ip')
-            if result != 'success':
-                return result, err_msg
-
-        if 'key' in expected:
-            result, err_msg = validate_argument(key, expected['key'], 'key')
-            if result != 'success':
-                return result, err_msg
-
-    elif expected['status'] == 'error':
-        result, err_msg = validate_argument(status, 'ERROR:', 'status')
-        if result != 'success':
-            return result, err_msg
-
-        message = response[1]
-        if 'message' in expected:
-            if re.match(expected['message'], message) is None:
-                return 'error', f"Invalid 'message': '{message}' received, '{expected['message']}' expected"
-    else:
-        raise Exception('Invalid expected status')
-
-    return result, err_msg
-
-
-def clean_agents_from_db():
-    """
-    Clean agents from DB
-    """
-    command = 'global sql DELETE FROM agent WHERE id != 0'
-    try:
-        query_wdb(command)
-    except Exception:
-        raise Exception('Unable to clean agents')
-
-
-def insert_agent_in_db(id=1, name="TestAgent", ip="any", registration_time=0, connection_status=0,
-                       disconnection_time=0):
-    """
-    Write agent in global.db
-    """
-    insert_command = f'global insert-agent {{"id":{id},"name":"{name}","ip":"{ip}","date_add":{registration_time}}}'
-    update_command = f'global sql UPDATE agent SET connection_status = "{connection_status}",\
-                       disconnection_time = "{disconnection_time}" WHERE id = {id};'
-    try:
-        query_wdb(insert_command)
-        query_wdb(update_command)
-    except Exception:
-        raise Exception(f'Unable to add agent {id}')
-
-
-@pytest.fixture(scope='function')
-def insert_pre_existent_agents(get_current_test_case, stop_authd_function):
-    agents = get_current_test_case.get('pre_existent_agents', [])
-    time_now = int(time.time())
-    try:
-        keys_file = open(CLIENT_KEYS_PATH, 'w')
-    except IOError as exception:
-        raise exception
-
-    clean_agents_from_db()
-
-    for agent in agents:
-        id = agent['id'] if 'id' in agent else '001'
-        name = agent['name'] if 'name' in agent else f'TestAgent{id}'
-        ip = agent['ip'] if 'ip' in agent else 'any'
-        key = agent['key'] if 'key' in agent else 'TopSecret'
-        connection_status = agent['connection_status'] if 'connection_status' in agent else 'never_connected'
-        if 'disconnection_time' in agent and 'delta' in agent['disconnection_time']:
-            disconnection_time = time_now + agent['disconnection_time']['delta']
-        elif 'disconnection_time' in agent and 'value' in agent['disconnection_time']:
-            disconnection_time = agent['disconnection_time']['value']
-        else:
-            disconnection_time = time_now
-        if 'registration_time' in agent and 'delta' in agent['registration_time']:
-            registration_time = time_now + agent['registration_time']['delta']
-        elif 'registration_time' in agent and 'value' in agent['registration_time']:
-            registration_time = agent['registration_time']['value']
-        else:
-            registration_time = time_now
-
-        # Write agent in client.keys
-        keys_file.write(f'{id} {name} {ip} {key}\n')
-
-        # Write agent in global.db
-        insert_agent_in_db(id, name, ip, registration_time, connection_status, disconnection_time)
-
-    keys_file.close()
+    for key in expected.keys():
+        if re.match(expected[key], response_dict[key]) is None:
+            return 'error', f"Invalid {key}: '{response_dict[key]}' received, '{expected[key]}' expected"
+    return 'success', ''
