+++ tests/integration/test_authd/conftest.py
@@ -1,15 +1,15 @@
 import pytest
+import time
 import os
 import yaml
+
 from wazuh_testing.tools import LOG_FILE_PATH, CLIENT_KEYS_PATH
+from wazuh_testing.wazuh_db import insert_agent_in_db, clean_agents_from_db
 from wazuh_testing.tools.file import truncate_file
 from wazuh_testing.tools.monitoring import FileMonitor, make_callback, AUTHD_DETECTOR_PREFIX
 from wazuh_testing.tools.configuration import write_wazuh_conf, get_wazuh_conf, set_section_wazuh_conf,\
                                               load_wazuh_configurations
-from wazuh_testing.tools.services import control_service, check_daemon_status, delete_dbs
-from wazuh_testing.tools.monitoring import QueueMonitor
-
-
+from wazuh_testing.tools.services import control_service
 from wazuh_testing.authd import DAEMON_NAME
 
 
@@ -149,3 +149,42 @@ def override_authd_force_conf(format_configuration):
 
     # Restore previous configuration
     write_wazuh_conf(backup_config)
+
+
+@pytest.fixture(scope='function')
+def insert_pre_existent_agents(get_current_test_case, stop_authd_function):
+    agents = get_current_test_case.get('pre_existent_agents', [])
+    time_now = int(time.time())
+    try:
+        keys_file = open(CLIENT_KEYS_PATH, 'w')
+    except IOError as exception:
+        raise exception
+
+    clean_agents_from_db()
+
+    for agent in agents:
+        id = agent['id'] if 'id' in agent else '001'
+        name = agent['name'] if 'name' in agent else f'TestAgent{id}'
+        ip = agent['ip'] if 'ip' in agent else 'any'
+        key = agent['key'] if 'key' in agent else 'TopSecret'
+        connection_status = agent['connection_status'] if 'connection_status' in agent else 'never_connected'
+        if 'disconnection_time' in agent and 'delta' in agent['disconnection_time']:
+            disconnection_time = time_now + agent['disconnection_time']['delta']
+        elif 'disconnection_time' in agent and 'value' in agent['disconnection_time']:
+            disconnection_time = agent['disconnection_time']['value']
+        else:
+            disconnection_time = time_now
+        if 'registration_time' in agent and 'delta' in agent['registration_time']:
+            registration_time = time_now + agent['registration_time']['delta']
+        elif 'registration_time' in agent and 'value' in agent['registration_time']:
+            registration_time = agent['registration_time']['value']
+        else:
+            registration_time = time_now
+
+        # Write agent in client.keys
+        keys_file.write(f'{id} {name} {ip} {key}\n')
+
+        # Write agent in global.db
+        insert_agent_in_db(id, name, ip, registration_time, connection_status, disconnection_time)
+
+    keys_file.close()
