--- tests/integration/test_authd/force_options/README.md
@@ -1,2 +1,0 @@
-# force options tests
-These tests aim to check the correct behavior of Authd under the different possible configurations of the force block.
