+++ tests/integration/test_authd/force_options/test_authd_force_options_invalid_config.py
@@ -45,6 +45,7 @@ os_version:
 
 tags:
     - enrollment
+    - authd
 '''
 import os
 import pytest
@@ -68,7 +69,7 @@ local_internal_options = {'authd.debug': '2'}
 tests = []
 test_case_ids = []
 for file in os.listdir(tests_path):
-    group_name = file.split(".")[0]
+    group_name = file.split('.')[0]
     file_tests = read_yaml(os.path.join(tests_path, file))
     tests = tests + file_tests
     test_case_ids = test_case_ids + [f"{group_name} {test_case['name']}" for test_case in file_tests]
@@ -135,9 +136,9 @@ def test_authd_force_options_invalid_config(get_current_test_case, configure_loc
 
     truncate_file(LOG_FILE_PATH)
     try:
-        control_service("restart", daemon=DAEMON_NAME)
+        control_service('restart', daemon=DAEMON_NAME)
     except Exception:
         pass
     else:
-        raise Exception("Authd started when it was expected to fail")
+        raise Exception('Authd started when it was expected to fail')
     validate_authd_logs(get_current_test_case.get('log', []))
